优化代码，增减功能
