+++ app/proguard-rules.pro
@@ -8,7 +8,7 @@
 # If your project uses WebView with JS, uncomment the following
 # and specify the fully qualified class name to the JavaScript interface
 # class:
-#-keepclassmembers class fqcn.of.javascript.interface.for.webview {
+#-keepclassmembers class fqcn.of.javascript.interface.for.webView {
 #   public *;
 #}
