+++ app/src/main/java/com/gitalive/isat/ui/home/HomeFragment.kt
@@ -1,0 +1,32 @@
+package com.gitalive.isat.ui.home
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.widget.Toast
+import androidx.fragment.app.Fragment
+import com.gitalive.isat.R
+import kotlinx.android.synthetic.main.webviewlayout.*
+
+class HomeFragment : Fragment() {
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        return inflater.inflate(R.layout.fragment_home, container, false)
+    }
+
+    override fun onActivityCreated(savedInstanceState: Bundle?) {
+        super.onActivityCreated(savedInstanceState)
+        val url = "http://txyminecraft.club:8123/"
+        webView.setInitialScale(100)
+        webView.settings.textZoom = 200
+        webView.loadUrl(url)
+        fab.setOnClickListener {
+            Toast.makeText(context, "页面已刷新", Toast.LENGTH_LONG).show()
+            webView.loadUrl(url)
+        }
+    }
+}
