+++ app/src/main/java/com/gitalive/isat/ui/web/WebFragment.kt
@@ -1,0 +1,50 @@
+package com.gitalive.isat.ui.web
+
+import android.annotation.SuppressLint
+import android.graphics.Bitmap
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.webkit.WebView
+import android.webkit.WebViewClient
+import android.widget.Toast
+import androidx.fragment.app.Fragment
+import com.gitalive.isat.R
+import kotlinx.android.synthetic.main.webviewlayout.*
+
+class WebFragment : Fragment() {
+
+    private lateinit var webViewModel: WebViewModel
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        webViewModel = WebViewModel(requireActivity().application)
+        return inflater.inflate(R.layout.fragment_web, container, false)
+    }
+
+    @SuppressLint("SetJavaScriptEnabled")
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onActivityCreated(savedInstanceState)
+        webView.settings.setSupportMultipleWindows(true)
+
+        webView.webViewClient = object : WebViewClient() {
+            override fun shouldOverrideUrlLoading(view: WebView?, url: String?): Boolean {
+                view?.loadUrl(url)
+                return true
+            }
+
+            override fun onPageStarted(view: WebView?, url: String?, favicon: Bitmap?) {
+                Toast.makeText(context, "页面加载中，请稍候", Toast.LENGTH_SHORT).show()
+            }
+        }
+        webView.loadUrl(webViewModel.getWeb().value)
+        fab.setOnClickListener {
+            Toast.makeText(context, "页面已刷新", Toast.LENGTH_LONG).show()
+            webView.loadUrl(webViewModel.getWeb().value)
+        }
+    }
+
+}
