+++ app/src/main/java/com/gitalive/isat/ui/web/WebViewModel.kt
@@ -1,0 +1,13 @@
+package com.gitalive.isat.ui.web
+
+import android.app.Application
+import androidx.lifecycle.AndroidViewModel
+import androidx.lifecycle.MutableLiveData
+
+class WebViewModel(application: Application) : AndroidViewModel(application) {
+    val url = "http://neverlag.gitee.io/minecraft-wiki/#/"
+
+    fun getWeb(): MutableLiveData<String> {
+        return MutableLiveData(url)
+    }
+}
