--- app/src/main/java/com/gitalive/myapplication/ui/about/AboutFragment.kt
@@ -1,20 +1,0 @@
-package com.gitalive.myapplication.ui.about
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import com.gitalive.myapplication.R
-
-class AboutFragment : Fragment() {
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_about, container, false)
-    }
-
-}
