+++ app/src/main/res/layout/fragment_bbs.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:configChanges="orientation|keyboardHidden|screenSize"
-    tools:context=".ui.home.HomeFragment">
+    tools:context=".ui.bbs.GalleryFragment">
 
     <com.gitalive.isat.ui.WebViewLayout
         android:layout_width="match_parent"
@@ -14,5 +14,4 @@
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
-
 </androidx.constraintlayout.widget.ConstraintLayout>
