--- app/src/main/res/menu/main.xml
@@ -1,40 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:tools="http://schemas.android.com/tools"
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <item
-        android:id="@+id/action_refresh"
-        android:enabled="true"
-        android:icon="@drawable/ic_refresh"
-        android:orderInCategory="1"
-        android:title="@string/action_refresh"
-        app:showAsAction="always"
-        tools:ignore="AlwaysShowAction" />
-
-    <item
-        android:id="@+id/action_back"
-        android:enabled="true"
-        android:icon="@drawable/ic_back"
-        android:orderInCategory="2"
-        android:title="@string/action_back"
-        app:showAsAction="always" />
-
-    <item
-        android:id="@+id/action_home"
-        android:enabled="true"
-        android:icon="@drawable/ic_home"
-        android:orderInCategory="3"
-        android:title="@string/action_home"
-        app:showAsAction="always" />
-
-    <item
-        android:id="@+id/action_forward"
-        android:enabled="true"
-        android:icon="@drawable/ic_forward"
-        android:orderInCategory="4"
-        android:title="@string/action_forward"
-        app:showAsAction="always" />
-
-
-</menu>
