--- app/src/main/res/menu/notice.xml
@@ -1,27 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:tools="http://schemas.android.com/tools"
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-    <item
-        android:id="@+id/action_feedback"
-        android:enabled="true"
-        android:icon="@drawable/ic_feedback"
-        android:orderInCategory="100"
-        android:title="@string/action_feedback"
-        app:showAsAction="always"
-        tools:ignore="AlwaysShowAction" />
-    <item
-        android:id="@+id/action_version"
-        android:enabled="true"
-        android:orderInCategory="100"
-        android:title="@string/action_version"
-        app:showAsAction="always"
-        tools:ignore="AlwaysShowAction" />
-    <item
-        android:id="@+id/action_refresh"
-        android:enabled="true"
-        android:icon="@android:drawable/stat_notify_sync"
-        android:orderInCategory="100"
-        android:title="@string/action_refresh"
-        app:showAsAction="always" />
-</menu>
