+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -3,38 +3,33 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/mobile_navigation"
-    app:startDestination="@+id/nav_home">
+    app:startDestination="@id/nav_home">
 
     <fragment
         android:id="@+id/nav_home"
-        android:name="com.gitalive.myapplication.ui.home.HomeFragment"
+        android:name="com.gitalive.isat.ui.home.HomeFragment"
         android:label="@string/menu_map"
         tools:layout="@layout/fragment_home" />
 
     <fragment
         android:id="@+id/nav_gallery"
-        android:name="com.gitalive.myapplication.ui.bbs.GalleryFragment"
+        android:name="com.gitalive.isat.ui.bbs.GalleryFragment"
         android:label="@string/menu_bbs"
-        tools:layout="@layout/fragment_gallery" />
+        tools:layout="@layout/fragment_bbs" />
 
     <fragment
         android:id="@+id/nav_slideshow"
-        android:name="com.gitalive.myapplication.ui.wiki.SlideshowFragment"
+        android:name="com.gitalive.isat.ui.wiki.SlideshowFragment"
         android:label="@string/menu_wiki"
-        tools:layout="@layout/fragment_slideshow" />
+        tools:layout="@layout/fragment_wiki" />
     <fragment
         android:id="@+id/nav_about"
-        android:name="com.gitalive.myapplication.ui.about.AboutFragment"
+        android:name="com.gitalive.isat.ui.about.AboutFragment"
         android:label="@string/menu_about"
         tools:layout="@layout/fragment_about" />
     <fragment
         android:id="@+id/nav_web"
-        android:name="com.gitalive.myapplication.ui.web.WebFragment"
+        android:name="com.gitalive.isat.ui.web.WebFragment"
         android:label="@string/menu_web"
-        tools:layout="@layout/web_fragment" />
-    <fragment
-        android:id="@+id/nav_bing"
-        android:name="com.gitalive.myapplication.ui.bing.BingFragment"
-        android:label="@string/menu_bing"
-        tools:layout="@layout/fragment_bing" />
+        tools:layout="@layout/fragment_web" />
 </navigation>
