+++ app/src/main/res/values-v21/strings.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="extend">extend</string>
+</resources>
