+++ app/src/main/res/values/strings.xml
@@ -1,26 +1,22 @@
 <resources>
-    <string name="app_name">QueDe Map</string>
+    <string name="app_name">ISaT</string>
     <string name="navigation_drawer_open">Open navigation drawer</string>
     <string name="navigation_drawer_close">Close navigation drawer</string>
-    <string name="nav_header_title">无限时空 Minecraft服务器</string>
-    <string name="nav_header_subtitle">A Minecraft Server of Infinite Time and Space</string>
     <string name="nav_header_desc">Navigation header</string>
-    <string name="action_version">3.0.1</string>
+    <string name="action_version">3.1.0</string>
     <string name="action_refresh">打开主页</string>
     <string name="action_forward">前往下一页</string>
     <string name="action_back">返回上一页</string>
 
-    <string name="menu_map">地图</string>
-    <string name="menu_bbs">论坛</string>
-    <string name="menu_wiki">维基</string>
-    <string name="menu_web">官网</string>
-    <string name="menu_bing">必应</string>
-    <string name="menu_about">关于</string>
-    <string name="author">App Made by Dell_G3</string>
+    <string name="menu_map">实时地图</string>
+    <string name="menu_bbs">中文论坛</string>
+    <string name="menu_wiki">游戏维基</string>
+    <string name="menu_web">问答服务</string>
+    <string name="menu_about">关于应用</string>
+    <string name="author">App Made by Git-a-Live</string>
     <string name="logo">Logo Designed by Mort</string>
-    <string name="version">Version: 3.0.1</string>
-    <string name="what_updated">3.0.1更新内容</string>
-    <string name="action_home">action_home</string>
-    <string name="action_feedback">action_feedback</string>
+    <string name="version">Version: 3.1.0</string>
+    <string name="extend">extend</string>
+    <string name="update_message">1.移除必应搜索；\n2.加入物品维基；\n3.修复部分bug。</string>
 
 </resources>
